Move calidation of argv's environment flag
The parsing is still done in duplicate, and I guess I blame lack of
generics for it.

That said, if I'd just make the config format bigger and more generic I
wouldn't need these specific types and it'd solve the problem a
different way. Added to the list of things to do.